move fetching to routeCard component
