+++ apps/site/assets/ts/tnm/components/RoutesSidebar.tsx
@@ -1,14 +1,11 @@
-import React, { ReactElement, useEffect } from "react";
+import React, { ReactElement } from "react";
 import { modeByV3ModeType } from "../../components/ModeFilter";
 import { Dispatch } from "../state";
-import { Alert, Mode, RouteWithStopsWithDirections, Stop } from "../../__v3api";
+import { Mode, RouteWithStopsWithDirections, Stop } from "../../__v3api";
 import RouteCard from "./RouteCard";
 import RouteSidebarPills from "./RouteSidebarPills";
 import ModeFilterContainer from "./ModeFilterContainer";
 import SidebarTitle from "./SidebarTitle";
-import { fetchJsonOrThrow, isFetchFailed } from "../../helpers/fetch-json";
-import { isHighSeverityOrHighPriority } from "../../models/alert";
-import useSWR from "swr";
 
 interface Props {
   data: RouteWithStopsWithDirections[];
@@ -24,57 +21,6 @@ interface FilterOptions {
   stopId: string | null;
   modes: Mode[];
 }
-
-export const fetchAlerts = async (routeId: string): Promise<Alert[]> => {
-  const results = await fetchJsonOrThrow<Alert[]>(
-    `/api/alerts?route_ids=${routeId}`
-  );
-
-  if (isFetchFailed(results)) {
-    // 404s here are a known failure mode, see finder_api.ex#get_trip_info
-    throw new Error(
-      `Failed to fetch Alert information: ${results.status} ${results.statusText}`
-    );
-  }
-
-  let alerts: Alert[] = [];
-  results.filter(isHighSeverityOrHighPriority);
-  for (let i = 0; i < results.length; i++) {
-    alerts.push(results[i] as Alert);
-  }
-  return alerts;
-};
-
-export const isUndefined = (arg: any): any => {
-  if (arg == undefined) {
-    return [];
-  }
-  return arg;
-};
-
-//fetchAlerts("Orange").then(function(val) {console.log(val[0] instanceof Alert)});
-// export const fetchAlerts = async (
-//   routeId: string,
-// ): Promise<Alert[]> => {
-//   let alerts: Alert[] = [];
-//   window.fetch &&
-//   window
-//     .fetch(`/api/alerts?route_ids=${routeId}`)
-//     .then(response => {
-//       if (response.ok) return response.json();
-//       throw new Error(response.statusText);
-//     })
-//     .then(result => {
-//         for (let alert in result) {
-//           alerts.push(JSON.parse(alert) as Alert);
-//         }
-//     console.log(alerts);
-
-//       return alerts;
-//     })
-//     console.log(alerts);
-//     return alerts;
-// };
 
 const filterDataByModes = (
   data: RouteWithStopsWithDirections[],
@@ -158,13 +104,6 @@ const RoutesSidebar = ({
     shouldFilterStopCards
   );
 
-  // const alerts = new Map<String, Alert[]>();
-
-  // useEffect(() => {
-  // filteredData.length > 0 ? filteredData.forEach(route => alerts.set(route.route.id, (useSWR<Alert[]>(`/api/alerts?route_ids=${"Orange"}`, fetchAlerts).data))) : [];
-  // console.log(alerts);
-  // });
-
   return (
     <div className="m-tnm-sidebar">
       <div className="m-tnm-sidebar__fixed-header">
@@ -190,12 +129,6 @@ const RoutesSidebar = ({
                   key={route.route.id}
                   route={route}
                   dispatch={dispatch}
-                  alerts={isUndefined(
-                    useSWR<Alert[]>(
-                      `/api/alerts?route_ids=${"Orange"}`,
-                      fetchAlerts
-                    ).data
-                  )}
                 />
               ))
             : emptyMessage}
